Remove `ServerApp` class.
