--- examples/src/main/java/org/spine3/examples/aggregate/server/ServerApp.java
@@ -1,162 +1,0 @@
-/*
- * Copyright 2016, TeamDev Ltd. All rights reserved.
- *
- * Redistribution and use in source and/or binary forms, with or without
- * modification, must retain the above copyright notice and the following
- * disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package org.spine3.examples.aggregate.server;
-
-import com.google.common.util.concurrent.MoreExecutors;
-import io.grpc.stub.StreamObserver;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.spine3.base.Command;
-import org.spine3.base.Response;
-import org.spine3.examples.aggregate.ClientApp;
-import org.spine3.server.BoundedContext;
-import org.spine3.server.command.CommandBus;
-import org.spine3.server.command.CommandStore;
-import org.spine3.server.event.EventBus;
-import org.spine3.server.event.EventStore;
-import org.spine3.server.event.EventSubscriber;
-import org.spine3.server.storage.StorageFactory;
-import org.spine3.server.storage.memory.InMemoryStorageFactory;
-
-import java.io.IOException;
-import java.util.List;
-
-/**
- * A sample application showing basic usage of the framework.
- *
- * @author Mikhail Mikhaylov
- * @author Alexander Litus
- */
-public class ServerApp implements AutoCloseable {
-
-    private final StorageFactory storageFactory;
-    private final BoundedContext boundedContext;
-    private final EventSubscriber eventLogger = new EventLogger();
-
-    /**
-     * Creates a new sample with the specified storage factory.
-     * @param storageFactory factory used to create and set up storages.
-     */
-    public ServerApp(StorageFactory storageFactory) {
-        this.storageFactory = storageFactory;
-
-        this.boundedContext = BoundedContext.newBuilder()
-                .setStorageFactory(storageFactory)
-                .setCommandBus(createCommandBus())
-                .setEventBus(createEventBus(storageFactory))
-                .build();
-    }
-
-    private static CommandBus createCommandBus() {
-        final CommandStore store = new CommandStore(InMemoryStorageFactory.getInstance().createCommandStorage());
-        final CommandBus commandBus = CommandBus.newInstance(store);
-        return commandBus;
-    }
-
-    private static EventBus createEventBus(StorageFactory storageFactory) {
-        final EventStore eventStore = EventStore.newBuilder()
-                .setStreamExecutor(MoreExecutors.directExecutor())
-                .setStorage(storageFactory.createEventStorage())
-                .setLogger(EventStore.log())
-                .build();
-        return EventBus.newInstance(eventStore);
-    }
-
-    /**
-     * The entry point of the sample.
-     */
-    public static void main(String[] args) throws Exception {
-        final StorageFactory factory = InMemoryStorageFactory.getInstance();
-
-        try (final ServerApp app = new ServerApp(factory)) {
-            app.execute();
-        } catch (IOException e) {
-            log().error("", e);
-        }
-    }
-
-    /**
-     * Executes the sample: generates some command requests and then the {@link BoundedContext} processes them.
-     */
-    public void execute() {
-        setUp();
-
-        // Generate test requests
-        final List<Command> requests = ClientApp.generateRequests();
-
-        final StreamObserver<Response> responseObserver = new StreamObserver<Response>() {
-            @Override
-            public void onNext(Response response) {
-            }
-
-            @Override
-            public void onError(Throwable throwable) {
-            }
-
-            @Override
-            public void onCompleted() {
-            }
-        };
-
-        // Process requests
-        for (Command request : requests) {
-            boundedContext.post(request, responseObserver);
-        }
-
-        log().info("All the requests were handled.");
-    }
-
-    /**
-     * Sets up the storage, initializes the bounded contexts, registers repositories, handlers etc.
-     */
-    public void setUp() {
-        // Register repository with the bounded context. This will register it in the CommandDispatcher too.
-        final OrderRepository repository = new OrderRepository(boundedContext);
-        repository.initStorage(storageFactory);
-
-        boundedContext.register(repository);
-
-        // Register event subscribers.
-        boundedContext.getEventBus().subscribe(eventLogger);
-    }
-
-    /**
-     * Tear down storages, unregister event handlers and close the bounded context.
-     */
-    @Override
-    public void close() throws Exception {
-        boundedContext.close();
-    }
-
-    public BoundedContext getBoundedContext() {
-        return boundedContext;
-    }
-
-    private static Logger log() {
-        return LogSingleton.INSTANCE.value;
-    }
-
-    private enum LogSingleton {
-        INSTANCE;
-        @SuppressWarnings("NonSerializableFieldInSerializableClass")
-        private final Logger value = LoggerFactory.getLogger(ServerApp.class);
-    }
-}
